--- QYWSR3/Commands/ClearCommand.cs
@@ -1,17 +1,0 @@
-﻿using System;
-
-internal class ClearCommand : CommandBase
-{
-    public ClearCommand(IHost host) : base(host)
-    {
-    }
-
-    public override string Name => "clear";
-
-    public override void Execute()
-    {
-        Host.Clear();
-    }
-}
-
-
