--- QYWSR3/Commands/ExitCommand.cs
@@ -1,17 +1,0 @@
-﻿using System;
-
-internal class ExitCommand : CommandBase
-{
-    public ExitCommand(IHost host) : base(host)
-    {
-    }
-
-    public override string Name => "exit";
-
-    public override void Execute()
-    {
-        Host.Exit();
-    }
-}
-
-
