--- QYWSR3/Commands/HelloCommand.cs
@@ -1,17 +1,0 @@
-﻿using System;
-
-internal class HelloCommand : CommandBase
-{
-    public HelloCommand(IHost host) : base(host)
-    {
-    }
-
-    public override string Name => "hello";
-
-    public override void Execute()
-    {
-        Host.WriteLine("Hello, User!");
-    }
-}
-
-
