--- QYWSR3/Consol/CommandRunner.cs
@@ -1,36 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-internal class CommandRunner : IHost
-{
-    private CommandLoader _cmdloader;
-    public CommandRunner()
-    {
-        _cmdloader = new CommandLoader(this);
-    }
-    public void Run()
-    {
-        while (true)
-        {
-            Console.Write("Command:");
-            string? userInput = Console.ReadLine();
-            if (!string.IsNullOrEmpty(userInput))
-            {
-                var command = _cmdloader.Commands
-                    .Where(command => command.Name == userInput)
-                    .FirstOrDefault();
-                if (command != null)
-                {
-                    command.Execute();
-                }
-                else
-                {
-                    Console.WriteLine($"This command doesn't exist: {userInput}");
-                }
-            }
-        }
-    }
-}
