--- QYWSR3/Consol/ICommand.cs
@@ -1,28 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Net;
-using System.Text;
-using System.Threading.Tasks;
-
-internal interface ICommand
-{
-    IHost Host { get; }
-    string Name { get; }
-    void Execute();
-
-}
-
-internal abstract class CommandBase : ICommand
-{
-    public IHost Host { get; }
-
-    protected CommandBase(IHost host)
-    {
-        Host = host;
-    }
-
-    public abstract string Name { get; }
-
-    public abstract void Execute();
-}
