--- QYWSR3/TODO list/TodoTask.cs
@@ -1,14 +1,0 @@
-﻿using System;
-
-internal record TodoTask
-{
-    public int Id { get; init; }
-    public string Description { get; init; }
-
-    public TodoTask(int id, string description)
-    {
-        Id = id;
-        Description = description;
-    }
-}
-
