--- QYWSR3/TODO list/todoList.json
@@ -1,1 +1,0 @@
-[{"Id":2,"Description":"hurka"},{"Id":3,"Description":"azaz"},{"Id":2,"Description":"task 435345"}]
