+++ QYWSR3/TodoListApplication/Commands/ExitCommand.cs
@@ -1,0 +1,20 @@
+﻿using System;
+using ToDoListApplication.TODOlist;
+
+namespace ToDoListApplication.Commands
+{
+    internal class ExitCommand : CommandBase
+    {
+        public ExitCommand(IHost host) : base(host)
+        {
+        }
+
+        public override string Name => "exit";
+
+        public override void Execute()
+        {
+            Host.Exit();
+        }
+    }
+}
+
