+++ QYWSR3/TodoListApplication/Commands/HelloCommand.cs
@@ -1,0 +1,19 @@
+﻿using System;
+using ToDoListApplication.TODOlist;
+
+namespace ToDoListApplication.Commands
+{
+    internal class HelloCommand : CommandBase
+    {
+        public HelloCommand(IHost host) : base(host)
+        {
+        }
+
+        public override string Name => "hello";
+
+        public override void Execute()
+        {
+            Host.WriteLine("Hello, User!");
+        }
+    }
+}
