+++ QYWSR3/TodoListApplication/Commands/UpdateCommand.cs
@@ -1,0 +1,26 @@
+﻿using System;
+using ToDoListApplication.TODOlist;
+
+namespace ToDoListApplication.Commands
+{
+    internal class UppdateCommand : CommandBase
+    {
+        private readonly TodoList todoList;
+
+        public UppdateCommand(IHost host, TodoList todoList) : base(host)
+        {
+            this.todoList = todoList;
+        }
+
+        public override string Name => "update";
+
+        public override async void Execute()
+        {
+            Host.WriteLine("Type in the name of the TODO list you want to update! It must be .json!");
+            Host.Write("Input:");
+            string fileNAME = Host.ReadLine();
+            await todoList.UpdateTODO(fileNAME);
+        }
+    }
+}
+
